Poor Man's Dedup of Large Microsoft Excel Data-Set
A friend's boss had asked her to edit a Microsoft Excel file containing about 40,000 rows manually (sigh!) to remove repeated patient records. I thought I would help out with a Python script to automate the process based on a .csv file version of the data. This is a "poor man's dedup" of patient records and my first experience to test things out on test cases where the data was confidential enough that I had to use dummy data to test my code. 

The funny thing is I was blissfully unaware of what Microsoft Excel is capable of then (I have since learned more on Excel, but not yet enough to deem myself an expert).

Basic concept:
1. Reads data from a .csv source file (datafile.csv; converted from a .xlsx/.xls).
2. Sorts patients by last name, then by first name and saves the data file (unresolved_data.csv).
3. Looks for repeated patient data (in unresolved_data.csv): retains most complete record and highest priority record based on ranking criteria speci